requirements.txt
boto3==1.24.40
botocore==1.27.40
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.1.3
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
